StringEdit

About

This package provides functions for calculating a String edit of minimum distance between two Strings (the notion of distance or cost here is synonymous with Levenshtein distance). That is, given two Strings, for each character, we can either associate the two according to some cost function or insert/delete a character from one of the Strings with some penalty (in this case, 1). Then we want to find a String edit of minimum distance.

Included is a naive implementation that takes exponential time, as well as a memoizing implementation that takes O(nm) time (where n and m are the lengths of each word). Additionally, the memoizing version allows us to recover a String edit of minimum distance.

cost :: Char -> Char -> Float

Calculate the cost of aligning two characters:

Vowel-to-vowel costs 0.5,

Vowel-to-consonant costs 1.2,

Consonant-to-consonant costs 0.6, and

Aligning the same characters costs 0.

Example Program

An example program run:

 $ runhaskell StringEdit.hs thenameofthegame theresmyname
 Minimum String Edit Cost: 7.0
 Minimum String Edit: 
 thenameofthegame
 ther  e smy name

Also note that the minimum String edit between A and B is the same as that between B and A, as expected.

 $ runhaskell StringEdit.hs ninakushukuru unamshukuru ninakushukuru
 Minimum String Edit Cost: 3.1
 Minimum String Edit: 
 ninakushukuru
  unam shukuru
 
 Minimum String Edit Cost: 3.1
 Minimum String Edit: 
  unam shukuru
 ninakushukuru

main :: IO ()

Calculates the minimum String edit between all adjacent pairs of Strings given.

Naive Implementation

naiveStringEditDist :: String -> String -> Float

Exponential-time implementation of the String edit distance algorithm.

Memoizing Implementation

memoStringEditDist :: String -> String -> State (Map (String, String) Float) Float

Memoizing implementation of naiveStringEditDist.

calcStringEditDist :: String -> String -> State (Map (String, String) Float) Float

O(nm). Calculate the String edit distance of two Strings using memoStringEditDist by building up the values of the memoization table from back to front.

stringEditDist :: String -> String -> Float

O(nm). Calculate the String edit distance of two Strings using calcStringEditDist.

String Edit Recovery

writeStringEdit :: String -> String -> Map (String, String) Float -> [(Maybe Char, Maybe Char)]

O(n). Trace back through a precomputed memoization table to recover the edit of minimum distance between two Strings.

stringEdit :: String -> String -> [(Maybe Char, Maybe Char)]

O(nm). Consruct the edit of minimum distance between two Strings using calcStringEditDist and writeStringEdit.

showStringEdit :: [(Maybe Char, Maybe Char)] -> String

Show a String edit.