StringEdit2

Documentation

cost :: Fractional a => Char -> Char -> a

Calculate the cost of aligning two characters: 1. Vowel-to-vowel costs 0.5, 2. Vowel-to-consonant costs 1.2, 3. Consonant-to-consonant costs 0.6, and 4. Aligning the same characters costs 0.

naiveStringEditDist :: String -> String -> Float

Exponential-time implementation of the string edit distance algorithm.

memoStringEditDist :: String -> String -> State (Map (String, String) Float) Float

Memoizing implementation of naiveStringEditDist.

calcStringEditDist :: String -> String -> State (Map (String, String) Float) Float

O(nm). Calculate the string edit distance of two Strings.

stringEditDist :: String -> String -> Float

O(nm). Calculate the string edit distance of two Strings.

writeStringEdit :: String -> String -> Map (String, String) Float -> [(Maybe Char, Maybe Char)]

O(n). Trace back through a precomputed memoization table to recover the edit of minimum distance between two Strings.

stringEdit :: String -> String -> [(Maybe Char, Maybe Char)]

O(nm). Consruct the edit of minimum distance between two Strings.

showStringEdit :: [(Maybe Char, Maybe Char)] -> String

Show a String edit.

main :: IO ()

Calculates the minimum string edit between all adjacent pairs of Strings given.